Add files via upload
The primary goal of this project is to delve into the effects of Ozempic on obese and
hypertensive patients aged 40-75. A key focus is on addressing endogeneity to estimate the
treatment effect accurately. Through exploratory data analysis, data merging, and the application of
advanced statistical techniques like Double lasso, RMSE, R-sq.